cleanup; promoted some code from GenerationTestBAse to SystemTestBase.
